feat(RPS-1438): refined Wandelscript types
