+++ examples/01_basic.py
@@ -2,6 +2,7 @@ import asyncio
 from pathlib import Path
 
 from nova import Nova, api
+from nova.types import Pose
 
 import wandelscript
 
@@ -24,7 +25,11 @@ async def main():
 
         robot_cell = await cell.get_robot_cell()
         run = wandelscript.run_file(
-            Path(__file__).parent / "01_basic.ws", cell=robot_cell, default_tcp=None, default_robot=None
+            Path(__file__).parent / "01_basic.ws",
+            cell=robot_cell,
+            default_tcp=None,
+            default_robot=None,
+            run_args={"pose_a": Pose((0, 0, 400, 0, 3.14, 0))},
         )
         print(run.program_run.execution_results)
