+++ wandelscript/examples.py
@@ -1,7 +1,6 @@
 import json
 from pathlib import Path
-
-from wandelscript import serializer
+from wandelscript.types import loads_store
 
 _EXAMPLE_DIR = Path(__file__).parent / "examples"
 
@@ -12,7 +11,7 @@ def read_examples():
     for a in _EXAMPLE_DIR.iterdir():
         if a.suffix == ".ws":
             code = a.open().read()
-            data = serializer.loads_store((_EXAMPLE_DIR / f"{a.stem}.json").open().read())
+            data = loads_store((_EXAMPLE_DIR / f"{a.stem}.json").open().read())
             try:
                 cell = json.load((_EXAMPLE_DIR / f"{a.stem}.cell.json").open())
             except FileNotFoundError:
