+++ wandelscript/tests/test_store.py
@@ -1,6 +1,7 @@
-from wandelscript import Store, serializer, types
+from wandelscript import Store
+from wandelscript.types import Record
 
 
 def test_store():
-    store = Store(init_vars={"a": 1, "b": 2, "r": types.Record(data={"x": 1, "y": 2})})
-    assert store.data_dict == {"a": 1, "b": 2, "r": serializer.Record(record={"x": 1, "y": 2})}
+    store = Store(init_vars={"a": 1, "b": 2, "r": Record(data={"x": 1, "y": 2})})
+    assert store.data_dict == {"a": 1, "b": 2, "r": Record(record={"x": 1, "y": 2})}
